+++ includes/providers/gam/api/class-targeting-keys.php
@@ -9,13 +9,13 @@ namespace Newspack_Ads\Providers\GAM\Api;
 
 use Newspack_Ads\Providers\GAM\Api;
 use Newspack_Ads\Providers\GAM\Api\Api_Object;
-use Google\AdsApi\AdManager\v202208\Statement;
-use Google\AdsApi\AdManager\v202208\String_ValueMapEntry;
-use Google\AdsApi\AdManager\v202208\TextValue;
-use Google\AdsApi\AdManager\v202208\SetValue;
-use Google\AdsApi\AdManager\v202208\CustomTargetingKey;
-use Google\AdsApi\AdManager\v202208\CustomTargetingValue;
-use Google\AdsApi\AdManager\v202208\ServiceFactory;
+use Google\AdsApi\AdManager\v202305\Statement;
+use Google\AdsApi\AdManager\v202305\String_ValueMapEntry;
+use Google\AdsApi\AdManager\v202305\TextValue;
+use Google\AdsApi\AdManager\v202305\SetValue;
+use Google\AdsApi\AdManager\v202305\CustomTargetingKey;
+use Google\AdsApi\AdManager\v202305\CustomTargetingValue;
+use Google\AdsApi\AdManager\v202305\ServiceFactory;
 
 /**
  * Newspack Ads GAM Default Targeting Keys